feat: Split example usage tests into seperate file, so that they are more maintainable
