+++ pyproject.toml
@@ -72,7 +72,15 @@ show_error_codes = true
 testpaths = ["tests"]
 asyncio_mode = "auto"
 asyncio_default_fixture_loop_scope = "session"
-addopts = "--cov=pytest_capsqlalchemy --cov-report=term-missing --cov-report=html --cov-report=xml --cov-append"
+addopts = """
+  --cov=pytest_capsqlalchemy
+  --cov-report=term-missing
+  --cov-report=html
+  --cov-report=xml
+  --cov-append
+  --ignore=tests/examples
+"""
+pytester_example_dir = "tests/examples"
 
 [tool.ruff]
 line-length = 120
