+++ tests/examples/test_missing_setup.py
@@ -1,0 +1,2 @@
+def test_capsqlalchemy_setup(capsqlalchemy):
+    pass
